wiper is now a shape
