+++ src/features/app/InputTabs.js
@@ -5,7 +5,6 @@ import {
 } from 'react-bootstrap'
 import { connect } from 'react-redux'
 import ShapeList from '../shapes/ShapeList'
-import Wiper from '../wiper/Wiper'
 import ThetaRho from '../theta_rho/ThetaRho'
 import { chooseInput } from '../app/appSlice'
 
@@ -19,10 +18,6 @@ class InputTabs extends Component {
        <Tabs defaultActiveKey="shapes" onSelect={this.handleSelect.bind(this)} id="input-tabs">
          <Tab eventKey="shapes" title="Shapes">
            <ShapeList />
-         </Tab>
-
-         <Tab eventKey="wiper" title="Wiper">
-           <Wiper />
          </Tab>
 
          <Tab eventKey="code" title="Import">
